Update data logging
Currently in the process of putting together a data storage directory
system that will work for schlieren and sensor data simultaneously.
Pressure and temperature logging now happens in parallel, and is
streamed to a common log file. Ignition tracking is done with a separate
log file in order to identify detonation events. This approach is an
improvement because it no longer clutters the disk with empty
directories for failed or aborted detonations. Pressure calibrations and
test conditions still need to be output to disk.